Split backup implementation into two parts to help implementing other backup engines.
