+++ src/Backups/IBackup.h
@@ -44,24 +44,27 @@ public:
     /// before the terminator. For example, list("", "") returns names of all the entries
     /// in the backup; and list("data/", "/") return kind of a list of folders and
     /// files stored in the "data/" directory inside the backup.
-    virtual Strings list(const String & prefix = "", const String & terminator = "/") const = 0;
+    virtual Strings listFiles(const String & prefix = "", const String & terminator = "/") const = 0;
 
     /// Checks if an entry with a specified name exists.
-    virtual bool exists(const String & name) const = 0;
+    virtual bool fileExists(const String & file_name) const = 0;
 
     /// Returns the size of the entry's data.
-    /// This function does the same as `read(name)->getSize()` but faster.
-    virtual size_t getSize(const String & name) const = 0;
+    /// This function does the same as `read(file_name)->getSize()` but faster.
+    virtual size_t getFileSize(const String & file_name) const = 0;
 
     /// Returns the checksum of the entry's data.
-    /// This function does the same as `read(name)->getCheckum()` but faster.
-    virtual UInt128 getChecksum(const String & name) const = 0;
+    /// This function does the same as `read(file_name)->getCheckum()` but faster.
+    virtual UInt128 getFileChecksum(const String & file_name) const = 0;
 
     /// Reads an entry from the backup.
-    virtual BackupEntryPtr read(const String & name) const = 0;
+    virtual BackupEntryPtr readFile(const String & file_name) const = 0;
 
     /// Puts a new entry to the backup.
-    virtual void write(const String & name, BackupEntryPtr entry) = 0;
+    virtual void addFile(const String & file_name, BackupEntryPtr entry) = 0;
+
+    /// Whether it's possible to add new entries to the backup in multiple threads.
+    virtual bool supportsWritingInMultipleThreads() const { return true; }
 
     /// Finalizes writing the backup, should be called after all entries have been successfully written.
     virtual void finalizeWriting() = 0;
